move personalisation input to separate file
